requirements.txt
# dependencias de python necesarias para fotoapp
pillow>=10.3.0
requests>=2.31.0
